move file and clean
