+++ Component/AnimationLoop/Sofa.Component.AnimationLoopConfig.cmake.in
@@ -1,0 +1,13 @@
+# CMake package configuration file for the @PROJECT_NAME@ module
+
+@PACKAGE_GUARD@
+@PACKAGE_INIT@
+
+find_package(Sofa.SimulationCore QUIET REQUIRED)
+find_package(Sofa.Component.Constraint.Lagrangian QUIET REQUIRED)
+
+if(NOT TARGET @PROJECT_NAME@)
+    include("${CMAKE_CURRENT_LIST_DIR}/@PROJECT_NAME@Targets.cmake")
+endif()
+
+check_required_components(@PROJECT_NAME@)
