+++ Component/AnimationLoop/src/sofa/component/animationloop/FreeMotionTask.h
@@ -66,4 +66,4 @@ private:
     bool m_parallelSolve {false };
 };
 
-}
+} // namespace sofa::component::animationloop
