+++ Component/AnimationLoop/src/sofa/component/animationloop/init.cpp
@@ -19,69 +19,33 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
-#include <SofaGeneralAnimationLoop/initSofaGeneralAnimationLoop.h>
+#include <sofa/component/animationloop/init.h>
 
-#include <sofa/helper/system/PluginManager.h>
-
-#include <sofa/core/ObjectFactory.h>
-using sofa::core::ObjectFactory;
-
-namespace sofa::component
+namespace sofa::component::animationloop
 {
-
-void initSofaGeneralAnimationLoop()
-{
-    static bool first = true;
-    if (first)
-    {
-        // msg_deprecated("SofaGeneralAnimationLoop") << "SofaGeneralAnimationLoop is deprecated. It will be removed at v23.06. You may use Sofa.Component.Mapping.MappedMatrix and Sofa.Component.AnimationLoop instead.";
-
-        sofa::helper::system::PluginManager::getInstance().loadPlugin("Sofa.Component.Mapping.MappedMatrix");
-        sofa::helper::system::PluginManager::getInstance().loadPlugin("Sofa.Component.AnimationLoop");
-
-        first = false;
-    }
-}
-
+    
 extern "C" {
-    SOFA_SOFAGENERALANIMATIONLOOP_API void initExternalModule();
-    SOFA_SOFAGENERALANIMATIONLOOP_API const char* getModuleName();
-    SOFA_SOFAGENERALANIMATIONLOOP_API const char* getModuleVersion();
-    SOFA_SOFAGENERALANIMATIONLOOP_API const char* getModuleLicense();
-    SOFA_SOFAGENERALANIMATIONLOOP_API const char* getModuleDescription();
-    SOFA_SOFAGENERALANIMATIONLOOP_API const char* getModuleComponentList();
+    SOFA_EXPORT_DYNAMIC_LIBRARY void initExternalModule();
+    SOFA_EXPORT_DYNAMIC_LIBRARY const char* getModuleName();
 }
 
 void initExternalModule()
 {
-    initSofaGeneralAnimationLoop();
+    static bool first = true;
+    if (first)
+    {
+        first = false;
+    }
 }
 
 const char* getModuleName()
 {
-    return sofa_tostring(SOFA_TARGET);
+    return MODULE_NAME;
 }
 
-const char* getModuleVersion()
+void init()
 {
-    return sofa_tostring(SOFAGENERALANIMATIONLOOP_VERSION);
+    initExternalModule();
 }
 
-const char* getModuleLicense()
-{
-    return "LGPL";
-}
-
-const char* getModuleDescription()
-{
-    return "This plugin contains contains features about General Animation Loop.";
-}
-
-const char* getModuleComponentList()
-{
-    /// string containing the names of the classes provided by the plugin
-    static std::string classes = ObjectFactory::getInstance()->listClassesFromTarget(sofa_tostring(SOFA_TARGET));
-    return classes.c_str();
-}
-
-} // namespace sofa::component
+} // namespace sofa::component::animationloop
