+++ modules/SofaConstraint/CMakeLists.txt
@@ -13,12 +13,9 @@ set(SOURCE_FILES
     )
 
 list(APPEND HEADER_FILES
-    ${SOFACONSTRAINT_SRC}/ConstraintAnimationLoop.h
     ${SOFACONSTRAINT_SRC}/ConstraintAttachBodyPerformer.h
     ${SOFACONSTRAINT_SRC}/ConstraintAttachBodyPerformer.inl
     ${SOFACONSTRAINT_SRC}/ContactIdentifier.h
-    ${SOFACONSTRAINT_SRC}/FreeMotionAnimationLoop.h
-    ${SOFACONSTRAINT_SRC}/FreeMotionTask.h
     ${SOFACONSTRAINT_SRC}/FrictionContact.h
     ${SOFACONSTRAINT_SRC}/FrictionContact.inl
     ${SOFACONSTRAINT_SRC}/LocalMinDistance.h
@@ -26,11 +23,8 @@ list(APPEND HEADER_FILES
     ${SOFACONSTRAINT_SRC}/StickContactConstraint.inl
     )
 list(APPEND SOURCE_FILES
-    ${SOFACONSTRAINT_SRC}/ConstraintAnimationLoop.cpp
     ${SOFACONSTRAINT_SRC}/ConstraintAttachBodyPerformer.cpp
     ${SOFACONSTRAINT_SRC}/ContactIdentifier.cpp
-    ${SOFACONSTRAINT_SRC}/FreeMotionAnimationLoop.cpp
-    ${SOFACONSTRAINT_SRC}/FreeMotionTask.cpp
     ${SOFACONSTRAINT_SRC}/FrictionContact.cpp
     ${SOFACONSTRAINT_SRC}/LocalMinDistance.cpp
     ${SOFACONSTRAINT_SRC}/StickContactConstraint.cpp
@@ -44,17 +38,19 @@ sofa_find_package(SofaUserInteraction REQUIRED)
 add_library(${PROJECT_NAME} SHARED ${HEADER_FILES} ${SOURCE_FILES})
 target_link_libraries(${PROJECT_NAME} PUBLIC SofaMeshCollision SofaImplicitOdeSolver SofaUserInteraction SofaBaseLinearSolver)
 
-message(WARNING "${PROJECT_NAME} module is being deprecated. It will be removed at v23.06. You may use Sofa.Component.Mapping.MappedMatrix, Sofa.Component.Constraint.Lagrangian.Model, Sofa.Component.Constraint.Lagrangian.Correction and Sofa.Component.Constraint.Lagrangian.Solver instead.")
+message(WARNING "${PROJECT_NAME} module is being deprecated. It will be removed at v23.06. You may use Sofa.Component.Mapping.MappedMatrix, Sofa.Component.Constraint.Lagrangian.Model, Sofa.Component.Constraint.Lagrangian.Correction, Sofa.Component.Constraint.Lagrangian.Solver and Sofa.Component.AnimationLoop instead.")
 
 # forward to the new sofang module
 sofa_find_package(Sofa.Component.Mapping.MappedMatrix REQUIRED)
 sofa_find_package(Sofa.Component.Constraint.Lagrangian.Model REQUIRED)
 sofa_find_package(Sofa.Component.Constraint.Lagrangian.Correction REQUIRED)
 sofa_find_package(Sofa.Component.Constraint.Lagrangian.Solver REQUIRED)
+sofa_find_package(Sofa.Component.AnimationLoop REQUIRED)
 target_link_libraries(${PROJECT_NAME} PUBLIC Sofa.Component.Mapping.MappedMatrix)
 target_link_libraries(${PROJECT_NAME} PUBLIC Sofa.Component.Constraint.Lagrangian.Model)
 target_link_libraries(${PROJECT_NAME} PUBLIC Sofa.Component.Constraint.Lagrangian.Correction)
 target_link_libraries(${PROJECT_NAME} PUBLIC Sofa.Component.Constraint.Lagrangian.Solver)
+target_link_libraries(${PROJECT_NAME} PUBLIC Sofa.Component.AnimationLoop)
 
 sofa_create_package_with_targets(
     PACKAGE_NAME ${PROJECT_NAME}
