+++ modules/SofaGeneralAnimationLoop/CMakeLists.txt
@@ -12,24 +12,15 @@ set(SOURCE_FILES
     ${SOFAGENERALANIMATIONLOOP_SRC}/initSofaGeneralAnimationLoop.cpp
     )
 
-list(APPEND HEADER_FILES
-    ${SOFAGENERALANIMATIONLOOP_SRC}/MultiStepAnimationLoop.h
-    ${SOFAGENERALANIMATIONLOOP_SRC}/MultiTagAnimationLoop.h)
-
-list(APPEND SOURCE_FILES
-    ${SOFAGENERALANIMATIONLOOP_SRC}/MultiStepAnimationLoop.cpp
-    ${SOFAGENERALANIMATIONLOOP_SRC}/MultiTagAnimationLoop.cpp)
-
-sofa_find_package(SofaBase REQUIRED) # SofaBaseLinearSolver
-
 add_library(${PROJECT_NAME} SHARED ${HEADER_FILES} ${SOURCE_FILES})
-target_link_libraries(${PROJECT_NAME} PUBLIC SofaSimulationCommon SofaBaseLinearSolver)
 
-message(WARNING "${PROJECT_NAME} module is being deprecated. It will be removed at v23.06. You may use Sofa.Component.Mapping.MappedMatrix instead.")
+message(WARNING "${PROJECT_NAME} module is deprecated. It will be removed at v23.06. Use Sofa.Component.Mapping.MappedMatrix and Sofa.Component.AnimationLoop instead.")
 
 # forward to the new sofang module
 sofa_find_package(Sofa.Component.Mapping.MappedMatrix REQUIRED)
+sofa_find_package(Sofa.Component.AnimationLoop REQUIRED)
 target_link_libraries(${PROJECT_NAME} PUBLIC Sofa.Component.Mapping.MappedMatrix)
+target_link_libraries(${PROJECT_NAME} PUBLIC Sofa.Component.AnimationLoop)
 
 sofa_create_package_with_targets(
     PACKAGE_NAME ${PROJECT_NAME}
