+++ modules/SofaGeneralAnimationLoop/SofaGeneralAnimationLoopConfig.cmake.in
@@ -5,13 +5,14 @@
 
 get_property(@PROJECT_NAME@_SENT_DEPRECATION_MESSAGE GLOBAL PROPERTY PROPERTY_@PROJECT_NAME@_SENT_DEPRECATION_MESSAGE SET)
 if(NOT @PROJECT_NAME@_SENT_DEPRECATION_MESSAGE)
-    message(WARNING "@PROJECT_NAME@ module is being deprecated. It will be removed at v23.06. You may use Sofa.Component.Mapping.MappedMatrix instead.")
+    message(WARNING "@PROJECT_NAME@ module is being deprecated. It will be removed at v23.06. Use Sofa.Component.Mapping.MappedMatrix and Sofa.Component.AnimationLoop instead.")
 endif()
 set_property(GLOBAL PROPERTY PROPERTY_@PROJECT_NAME@_SENT_DEPRECATION_MESSAGE TRUE)
 
 find_package(SofaBase QUIET REQUIRED)
 find_package(SofaSimulationCommon QUIET REQUIRED)
 find_package(Sofa.Component.Mapping.MappedMatrix QUIET REQUIRED)
+find_package(Sofa.Component.AnimationLoop QUIET REQUIRED)
 
 if(NOT TARGET @PROJECT_NAME@)
     include("${CMAKE_CURRENT_LIST_DIR}/@PROJECT_NAME@Targets.cmake")
